Cleaned up/reworked to batter match new workflow
`cli.js` is no longer where the majority of the work is done. It's simply a wrapper to configure and call `index.js`.

The result of calling `index.js` is still just logged to the console for now.

`index.js` now passes arguments through to the `builder` and works through the specified `pipeline`.

`level.properties` and `level.directions_allowed` are deleted because they are not useful to anything outside `index.js`.