Adding new resources for the reworked project structure
There is still work to be done here, but there is now a much cleaner seperation between the three parts of the project:

1. Plumbing code (`arguments.js`, `builder.js`, `cli.js`, and `index.js`)
2. Config code   (everything under `recipies/DEFAULT/config`)
3. Delegate code (everything under `recipies/DEFAULT/delegates`)

The idea behind this structure is:

1. Clients will not touch the Plumbing code
2. Clients will often update Config code
3. Clients will rarely update Delegate code

The code under `recipies/DEFAULT` is the default, built-in, support `volor` offers. The plan is to add support for loading external Config and Delegate code, with `DEFAULT` filling in any gaps.