Switched naming from "level" to "recipe"
"level" doesn't really match what the variable is meant or responsible for.

"recipe" matches what this thing is meant to be and what it actually does. A recipe is a bunch or rules and boundaries describing how content should be generated. Right now it's being used to generate level segments for a platformer game, but it can (and will, very soon) do much more than that.