Added `meta` section
Added a `meta` section with the `reserved` property and a `render_order` property.

`meta` isn't being used everywhere yet, but the idea is that information about the config will be placed here.

For example, most configs now have a `reserved` property. This is to allow delegate handlers to be smarter about what data in the config they process, and what data they ignore.

In this config, I also need to specify the order in which I want landscape elements rendered. This is important to guarantee consistency in the landscape output.